LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__host.header.https-expected.txt


PASS __Host: Secure origin: Does not set 'Path=/;'
FAIL __HoSt: Secure origin: Does not set 'Path=/;' assert_equals: expected (undefined) undefined but got (string) "0.8300989000957126"
PASS __Host: Secure origin: Does set 'Secure; Path=/;'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; '
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; ' assert_equals: expected (undefined) undefined but got (string) "0.9952085028334829"
PASS __Host: Secure origin: Does not set 'Path=/;MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Path=/;MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "0.21375505748168466"
PASS __Host: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; MaxAge=10' assert_equals: expected (undefined) undefined but got (string) "0.4369157037414726"
PASS __Host: Secure origin: Does not set 'Path=/;HttpOnly'
FAIL __HoSt: Secure origin: Does not set 'Path=/;HttpOnly' assert_equals: expected (undefined) undefined but got (string) "0.717933626017238"
PASS __Host: Secure origin: Does set 'Secure; Path=/;HttpOnly'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;HttpOnly'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; HttpOnly'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; HttpOnly' assert_equals: expected (undefined) undefined but got (string) "0.2167002660427767"
PASS __Host: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py' assert_equals: expected (undefined) undefined but got (string) "0.20346703720585624"
